refactor: Move type to domain
