+++ src/components/map-details-chart-adapters.tsx
@@ -6,10 +6,8 @@ import { Loading } from "src/components/hint";
 import { NetworkCostTrendChart } from "src/components/network-cost-trend-chart";
 import { PowerStabilityChart } from "src/components/power-stability-chart";
 import { TariffsTrendChart } from "src/components/tariffs-trend-chart";
-import {
-  SunshineIndicator,
-  useQueryStateSunshineMap,
-} from "src/domain/query-states";
+import { useQueryStateSunshineMap } from "src/domain/query-states";
+import { SunshineIndicator } from "src/domain/data";
 import {
   useEnergyTariffsQuery,
   useNetTariffsQuery,
